Fix compiler warnings on warning level 4 for msvc.

+++ sabre/include/sabre/Type_Interner.h
@@ -832,7 +832,7 @@ namespace sabre
 				return t->struct_type.fields[index].type;
 			return nullptr;
 		case Type::KIND_ARRAY:
-			if (index < t->array.count)
+			if ((int64_t)index < t->array.count)
 				return t->array.base;
 			return nullptr;
 		case Type::KIND_ENUM:
